Service로 로직 옮기기
